requirements.txt
aiomysql==0.1.1
amqp==5.1.1
anyio==3.6.1
argcomplete==2.0.0
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
black==22.6.0
celery==5.2.7
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
datamodel-code-generator==0.11.19
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.0
fastapi-code-generator==0.3.5
genson==1.2.2
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
inflect==5.6.2
isodate==0.6.1
isort==5.10.1
Jinja2==3.1.2
jsonschema==3.2.0
kombu==5.2.4
MarkupSafe==2.1.1
mypy==0.961
mypy-extensions==0.4.3
openapi-schema-validator==0.1.6
openapi-spec-validator==0.3.3
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
prance==0.21.8.0
prompt-toolkit==3.0.30
pyasn1==0.4.8
pydantic==1.9.1
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.16.1
PySnooper==0.5.0
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
rfc3986==1.5.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
semver==2.13.0
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
sqlalchemy2-stubs==0.0.2a24
starlette==0.19.1
stringcase==1.2.0
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.10
uvicorn==0.18.2
uvloop==0.16.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
